refactor: move src to legacy and create re-export index for compatibility
