+++ src/index.ts
@@ -1,7 +1,7 @@
-import useNotification from './hooks/useNotification';
-import Notice from './Notice';
-import type { NotificationAPI, NotificationConfig } from './hooks/useNotification';
-import NotificationProvider from './NotificationProvider';
+import useNotification from './legacy/hooks/useNotification';
+import Notice from './legacy/Notice';
+import type { NotificationAPI, NotificationConfig } from './legacy/hooks/useNotification';
+import NotificationProvider from './legacy/NotificationProvider';
 
 export { useNotification, Notice, NotificationProvider };
 export type { NotificationAPI, NotificationConfig };
