+++ src/legacy/Notice.tsx
