+++ src/legacy/interface.ts
